requirements.txt
numpy
RPi.GPIO
tflite
tflite-support>=0.3.3
tflite-runtime
opencv-python
picamera2
gpiozero
adafruit-blinka
adafruit-circuitpython-servokit